fix(wasm): fail closed on an unrecognized metadata-filter condition type
evaluate_condition's default match arm returned true for any "type" it
did not recognize, so a typo'd or unsupported condition (e.g. "eqals"
instead of "eq") silently matched every payload instead of being
rejected. search_with_filter is the only caller and passes filters
straight through with no validation of "type" beyond JSON shape, so a
malformed filter would quietly return unfiltered results through the
browser SDK. Flip the default to false and add a regression test.

Moves the inline test module to a sibling filter_tests.rs, matching the
crate's established convention and the shrink-only inline-test-module
baseline (adding to the inline block would have grown frozen debt);
scripts/inline-tests-baseline.txt drops the now-empty entry.
