+++ crates/velesdb-wasm/src/filter.rs
@@ -76,7 +76,9 @@ pub fn evaluate_condition(payload: &Value, condition: &Value) -> bool {
         "not" => condition
             .get("condition")
             .is_none_or(|c| !evaluate_condition(payload, c)),
-        _ => true,
+        // Fail closed: an unrecognized/misspelled "type" must not silently
+        // match everything, or a filter typo would leak unfiltered results.
+        _ => false,
     }
 }
 
@@ -126,106 +128,5 @@ pub fn get_nested_field<'a>(payload: &'a Value, field: &str) -> Option<&'a Value
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    #[test]
-    fn test_filter_eq() {
-        let payload = json!({"category": "tech"});
-        let filter = json!({
-            "condition": {
-                "type": "eq",
-                "field": "category",
-                "value": "tech"
-            }
-        });
-        assert!(matches_filter(&payload, &filter));
-    }
-
-    #[test]
-    fn test_filter_neq() {
-        let payload = json!({"category": "tech"});
-        let filter = json!({
-            "condition": {
-                "type": "neq",
-                "field": "category",
-                "value": "sports"
-            }
-        });
-        assert!(matches_filter(&payload, &filter));
-    }
-
-    #[test]
-    fn test_filter_gt() {
-        let payload = json!({"score": 85.0});
-        let filter = json!({
-            "condition": {
-                "type": "gt",
-                "field": "score",
-                "value": 80.0
-            }
-        });
-        assert!(matches_filter(&payload, &filter));
-    }
-
-    #[test]
-    fn test_filter_and() {
-        let payload = json!({"category": "tech", "score": 90.0});
-        let filter = json!({
-            "condition": {
-                "type": "and",
-                "conditions": [
-                    {"type": "eq", "field": "category", "value": "tech"},
-                    {"type": "gt", "field": "score", "value": 80.0}
-                ]
-            }
-        });
-        assert!(matches_filter(&payload, &filter));
-    }
-
-    #[test]
-    fn test_filter_or() {
-        let payload = json!({"category": "sports"});
-        let filter = json!({
-            "condition": {
-                "type": "or",
-                "conditions": [
-                    {"type": "eq", "field": "category", "value": "tech"},
-                    {"type": "eq", "field": "category", "value": "sports"}
-                ]
-            }
-        });
-        assert!(matches_filter(&payload, &filter));
-    }
-
-    #[test]
-    fn test_filter_not() {
-        let payload = json!({"category": "tech"});
-        let filter = json!({
-            "condition": {
-                "type": "not",
-                "condition": {
-                    "type": "eq",
-                    "field": "category",
-                    "value": "sports"
-                }
-            }
-        });
-        assert!(matches_filter(&payload, &filter));
-    }
-
-    #[test]
-    fn test_nested_field() {
-        let payload = json!({"user": {"profile": {"name": "John"}}});
-        let value = get_nested_field(&payload, "user.profile.name");
-        assert_eq!(value, Some(&json!("John")));
-    }
-
-    #[test]
-    fn test_no_filter_matches_all() {
-        let payload = json!({"anything": "value"});
-        let filter = json!({});
-        assert!(matches_filter(&payload, &filter));
-    }
-}
+#[path = "filter_tests.rs"]
+mod tests;
